Main final

+++ srcs_bot/balatro/command/deselect.cpp
@@ -6,7 +6,7 @@
 /*   By: mateo <mateo@42angouleme.fr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/11 18:07:16 by mateo             #+#    #+#             */
-/*   Updated: 2025/05/12 12:34:41 by mateo            ###   ########.fr       */
+/*   Updated: 2025/05/20 18:01:17 by mbico            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 void	Table::insertPlayCardinHand()
 {
-	for (int i = 0; i < _playHand.size(); i ++)
+	for (int i = 0; i < (int)_playHand.size(); i ++)
 	{
 		_hand.push_back(_playHand[i]);	
 	}
